clippy: clean across the kernel's feature combos and the userspace workspace
Two of these were deny-level lints that made `cargo clippy` on programs/ fail
outright, so nobody had been seeing the rest.

`top`'s wait loop never looped: every path broke, so any keystroke forced an
immediate redraw. That contradicts the comment directly above it -- "a key that
is not a command is not worth a redraw, so the wait resumes with whatever is
left of the interval" -- and the cause was `handle_key` returning a bare bool,
which cannot distinguish "changed an option" from "ignored". It returns a
three-way `Key` now and the loop does what the comment always claimed.

`auxvtest`'s `main` dereferences a raw pointer through `read_auxv`. It is the C
entry point, so its signature is fixed by the ABI and it cannot be marked
unsafe; annotated, with `read_auxv` carrying the real contract.

edos_lib's socket wrappers returned `Result<_, ()>`, discarding the errno the
kernel had just given them: a caller could not tell a refused connection from
an unreachable host. They carry `NetError` now. Three call sites needed a
touch, not the nineteen crates that use the API.

The rest is mechanical -- needless returns, manual prefix strips, range loops
that index -- across 53 files, plus a `too many arguments` in the NVMe submit
path that became a `Transfer` struct, which reads better at the call sites
anyway since the four fields travel together through the split loop.

+++ programs/edos-terminal/src/main.rs
@@ -152,8 +152,8 @@ fn main() {
                         match event.code {
                             0 => {
                                 terminal.on_mouse_button(
-                                    event.x as i32,
-                                    event.y as i32,
+                                    event.x,
+                                    event.y,
                                     event.data != 0,
                                 );
                             }
@@ -165,7 +165,7 @@ fn main() {
                         // Only a selection drag changes the picture; the
                         // pointer merely crossing the window does not.
                         if terminal.is_selecting() {
-                            terminal.on_mouse_move(event.x as i32, event.y as i32);
+                            terminal.on_mouse_move(event.x, event.y);
                             content_changed = true;
                         }
                     }
